+++ packages/restate-sdk/src/endpoint/fetch_endpoint.ts
@@ -24,7 +24,7 @@ import type {
 } from "../endpoint.js";
 import { GenericHandler } from "./handlers/generic.js";
 import { fetcher } from "./handlers/fetch.js";
-import type { ProtocolMode } from "../types/discovery.js";
+import { ProtocolMode } from "../types/discovery.js";
 
 /**
  * Generic Fetch encapsulates all the Restate services served by this endpoint.
@@ -39,13 +39,23 @@ import type { ProtocolMode } from "../types/discovery.js";
  *   .endpoint()
  *   .bind(myService)
  *   .handler();
+ * @example
+ * A typical bidirectional handler (must be served over http2) would look like this:
+ * ```
+ * import * as restate from "@restatedev/restate-sdk/fetch";
+ *
+ * export default restate
+ *   .endpoint()
+ *   .bind(myService)
+ *   .enableHttp2()
+ *   .handler();
  */
 export interface FetchEndpoint extends RestateEndpointBase<FetchEndpoint> {
   handler(): { fetch: (request: Request) => Promise<Response> };
 }
 
 export class FetchEndpointImpl implements FetchEndpoint {
-  constructor(private readonly protocolMode: ProtocolMode) {}
+  constructor(private protocolMode: ProtocolMode) {}
   private builder: EndpointBuilder = new EndpointBuilder();
 
   public get keySet(): KeySetV1 | undefined {
@@ -77,6 +87,11 @@ export class FetchEndpointImpl implements FetchEndpoint {
 
   public withIdentityV1(...keys: string[]): FetchEndpoint {
     this.builder.withIdentityV1(...keys);
+    return this;
+  }
+
+  public enableHttp2(): FetchEndpoint {
+    this.protocolMode = ProtocolMode.BIDI_STREAM;
     return this;
   }
