+++ packages/restate-sdk/src/fetch.ts
@@ -19,6 +19,7 @@ import { ProtocolMode } from "./types/discovery.js";
 
 /**
  * Create a new {@link RestateEndpoint} in request response protocol mode.
+ * Bidirectional mode (must be served over http2) can be enabled with .enableHttp2()
  */
 export function endpoint(): FetchEndpoint {
   return new FetchEndpointImpl(ProtocolMode.REQUEST_RESPONSE);
